Refactor (soon to be) shared behaviour from form models
This commit refactors the behaviour (mostly validations) that will be
shared between the database backed PlacementRequest and the form models.
It also refactors the relevent specs we'll want to share into shared
examples.
