+++ spec/support/background_check_shared_examples.rb
@@ -1,0 +1,5 @@
+shared_examples 'a background check' do
+  context 'attributes' do
+    it { is_expected.to respond_to :has_dbs_check }
+  end
+end
